Added google-login route
This route calls the googleLogin function, which will detect if a user is already registered, if thats the case then the user will be automatically logged in. If the user isnt registered, then an account will be created for him, with a default password of 1234, however, he wont be able to login the normal way with that password.